requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.5
Werkzeug==3.0.1
python-socketio==5.10.0
eventlet==0.33.3
pycryptodome==3.19.0
pyaudio==0.2.14
